Move multiple args out of server_args
